requirements.txt
python-snippets
openai
zhipuai
requests
pandas